requirements.txt
Django==5.2.1
django-crispy-forms==2.4
crispy-bootstrap5==2025.4
django-rest-framework==3.16.0